Rework the expand pass a bit.
Instead of searching for a pixel with nonzero alpha and returning it, we blend
neighboring pixels based on their alpha, which handles having a very transparent
(but not completely) neighboring pixel more robustly.  We also blend under the
pixel itself, so if the pixel has 1 alpha we won't use any neighboring color, if
it has 0.5 then we'll fill in the other 50% with neighboring color, and if it has
0.01 alpha we'll fill 99% with neighboring color.

This helps deal with premultiplication accuracy problems when we unpremultiply
later.  If we leave a pixel with very little alpha and then unpremultiply it, we'll
lose too much precision and end up with pixels with color inaccurate enough to
be noticeable.

This also reduces sampling from all 8 directions to just 4, which seems to be enough
and reduces texture sampling.  This is only run on the mosaic-resolution texture, so
this probably doesn't matter very much.