Remove PremultiplyShader.  It's not documented anywhere at all, but Unity RenderTextures are already premultiplied. Improve the way we figure out if we need to recreate textures. Make the expand pass have a texture in the textures (now Passes) array.  This is simpler and lets it be visualized in the inspector more easily.
Now rendering an object and setting Alpha to 0 to turn off the mosaic looks exactly the same as without the script, even with MSAA.  Before, the antialiasing wasn't quite correct.

This also reduces the number of expand passes to just 1.  More than one pass never seems to be necessary with the updated expand logic.