Make transform and scale anchoring more separate.
You can turn transform anchoring on without scale anchoring.  The mosaic will move to
follow the anchor, but the mosaic size will stay fixed.

You can turn on scale anchoring without transform anchoring.  The mosaic will resize itself
to match the distance to the anchor, but other than movement due to the number of blocks
changing, the mosaic will still stay fixed on screen.

You can turn both on, which will make the mosaic follow the object, as well as scale based
on its distance.

We'll now retain the mosaic offset when transform anchoring is turned on, so it won't
jump when it's turned on and off.  Instead, the mosaic will stay fixed and start following
the anchor from there.