Add a heavier demo scene using the Unity-chan asset to show more advanced use.
Remove the outline example from the demo project.  It's meant to show that shaders like
toon outlines can be overlaid on top of the mosaic and not mosaiced, but I think it's confusing
and makes it look like cell shading is a feature of the mosaic itself, which it isn't.
Instead, this new demo scene shows this.